Version 1.0
This is an online Auction system which has functionality of an actual auction system with live bidding.

This project contains files that have been  licensed under the owner's terms. These files are:

1) Twitter Bootstrap
https://getbootstrap.com/

2) Font Awesome Icons
https://fontawesome.com/

3) Images
The Images used in this project is just for the demo purpose and are owned by different owners like Flipkart, Amazon and other sources.

4) PHP Mailer
https://github.com/PHPMailer/PHPMailer